jetson/requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3
pyserial>=3.5
python-dotenv>=0.19.0 